Created base EmailTagInput.jsx along with the NotificationContacts component. Updated AsrSpecificNotificationDetails.jsx to use the EmailTagInput.jsx.
