+++ Tombolo/client-reactjs/src/components/common/Monitoring/NotificationContacts.jsx
@@ -1,0 +1,28 @@
+import { useSelector } from 'react-redux';
+import { Card, Form } from 'antd';
+import React from 'react';
+import EmailTagInput from '../EmailTagInput';
+import AsrSpecificNotificationsDetails from './AsrSpecificNotificationDetails';
+
+export default function NotificationContacts({ form, children }) {
+  // Redux
+  const {
+    applicationReducer: {
+      application: { applicationId },
+      integrations,
+    },
+  } = useSelector((state) => state);
+  const asrIntegration = integrations.some(
+    (integration) => integration.name === 'ASR' && integration.application_id === applicationId
+  );
+
+  return (
+    <Card>
+      <Form form={form} layout="vertical">
+        {children}
+        <EmailTagInput label="Primary Contact(s)" name="primaryContacts" required={true} />
+        {asrIntegration && <AsrSpecificNotificationsDetails />}
+      </Form>
+    </Card>
+  );
+}
